// 2022/src/Day02/enums.ts
export enum Opponent {
  rock = 'A',
  paper = 'B',
  scissors = 'C',
}

export enum Player {
  rock = 'X',
  paper = 'Y',
  scissors = 'Z',
}

export enum Points {
  rock = 1,
  paper = 2,
  scissors = 3,

  lose = 0,
  draw = 3,
  win = 6,
}

export enum DesiredOutcome {
  lose = 'X',
  draw = 'Y',
  win = 'Z',
}